BLM, CalTrans, Exclude Tracking

+++ src/scripts/_3_CNRA_Points.py
@@ -8,40 +8,40 @@
 # Version: 1.0.0
 # Date Created: Jan 24, 2024
 """
+import os
+import time
+import datetime
 import arcpy
 from ._2b_assign_domains import AssignDomains
 from ._2f_calculate_category import Category
+# add 2j standardize domains
 from ._2m_counts_to_mas import CountsToMAS
 from ._7b_enrichments_pts import enrich_points
 from ._2j_standardize_domains import StandardizeDomains
-from .utils import init_gdb, delete_scratch_files, runner
-# import os, 
-import time
-import datetime
+from .utils import init_gdb, delete_scratch_files
 
 date_id = datetime.datetime.now().strftime("%Y-%m-%d").replace("-", "")  # like 20221216
-
-original_gdb, workspace, scratch_workspace = init_gdb()
-runner(workspace, scratch_workspace)
+# TODO add print steps, rename variables
+workspace, scratch_workspace = init_gdb()
 
 def CNRA_pts_Model(
-    input_pt_fc, Activity_Table, Project_Poly, WFR_TF_Template, output_pt_enriched
+    input_pt_fc, Activity_Table, Project_Poly, WFR_TF_Template, output_pt_enriched,
+    delete_scratch=True
 ):
     start = time.time()
     print(f"Start Time {time.ctime()}")
     
     # Model Environment settings
     with arcpy.EnvManager(
+        workspace=workspace,
+        scratchWorkspace=scratch_workspace, 
         outputCoordinateSystem= arcpy.SpatialReference("NAD 1983 California (Teale) Albers (Meters)"), #WKID 3310
         cartographicCoordinateSystem=arcpy.SpatialReference("NAD 1983 California (Teale) Albers (Meters)"), #WKID 3310
-        extent="""450000, -374900, 540100, -604500,
-                  DATUM["NAD 1983 California (Teale) Albers (Meters)"]""",
+        extent="xmin=-374900, ymin=-604500, xmax=540100, ymax=450000, spatial_reference='NAD 1983 California (Teale) Albers (Meters)'", 
         preserveGlobalIds=True, 
         qualifiedFieldNames=False, 
-        scratchWorkspace=scratch_workspace, 
         transferDomains=False, 
-        transferGDBAttributeProperties=True, 
-        workspace=workspace,
+        transferGDBAttributeProperties=False, 
         overwriteOutput = True,
     ):
 
@@ -53,7 +53,7 @@ def CNRA_pts_Model(
         arcpy.management.CopyFeatures(input_pt_fc, Input_Features_1)
         # Input_Features_3 = arcpy.DefineProjection_management(Input_Features_2, "NAD 1983 California (Teale) Albers (Meters)") # WKID 3310
         
-        # ## Attribute Validation
+        # ## Attribute Validation (if needed)
         # # Process: Calculate Project ID User Field (4) (Calculate Field) (management)
         # Input_Features_2a = arcpy.management.CalculateField(
         #     in_table=Input_Features_1,
@@ -730,15 +730,12 @@ def CNRA_pts_Model(
         # Process: 2b Assign Domains (5) (2b Assign Domains) 
         output_pt_enriched = AssignDomains(in_table=output_pt_enriched)
 
-        print("Deleting Scratch Files")
-        delete_scratch_files(
-            gdb=scratch_workspace, delete_fc="yes", delete_table="yes", delete_ds="yes"
-        )
+        if delete_scratch: delete_scratch_files(
+                gdb=scratch_workspace, delete_fc="yes", delete_table="yes", delete_ds="yes"
+            )
 
         end = time.time()
         print(f"Time Elapsed: {(end-start)/60} minutes")
 
     return output_pt_enriched 
 
-# if __name__ == "__main__":
-#         CNRA_pts_Model()
